pyproject.toml
[project]
name = "nlp-be"
version = "0.1.0"
description = "NLP Backend Project"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    # --- [FastAPI & Server] ---
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart",
    "requests",
    # --- [DB & Auth] ---
    "sqlalchemy>=2.0.0",
    "redis>=5.0.0",
    "passlib[bcrypt]",
    "PyJWT",
    # --- [AI & LLM] ---
    "langchain",
    "langchain-openai",
    "langchain-community",
    "langchain-chroma",
    "openai",
    "groq",
    # --- [Vector DB & Embeddings] ---
    "chromadb>=0.4.24",
    "sentence-transformers",
    "numpy<2.0.0", # 호환성을 위해 2.0 미만 유지
    # --- [Data Processing & Utils] ---
    "pydantic>=2.0.0", # 최신 버전 (v2) 사용
    "python-dotenv",
    "beautifulsoup4",
    "pdfplumber",
    "pillow",
    "pytesseract",
    "playwright",
    "jinja2",
    # --- [Frontend (Optional)] ---
    "streamlit",
    "argon2-cffi>=25.1.0",
]

[tool.uv]
dev-dependencies = []
